ai-api/requirements.txt
grpcio==1.48.2
grpcio-tools==1.48.2
numpy==1.23.5
matplotlib==3.6.2